Added the log equation.

+++ src/Equation.java
@@ -1,3 +1,3 @@
 abstract class Equation {
-    abstract float calculate(float x);
+    abstract float calculate(double x);
 }
